All Star Code Challenge #31
Walter has a doctor's appointment and has to leave Jesse in charge of preparing their next "cook". He left Jesse an array of Instruction objects, which contain solutions, amounts, and the appropriate scientific instrument to use for each step in the "cook". The order of the instructions must be carried out in the EXACT order they're given. However, Jesse doesn't understand JavaScript and just needs things simplified!
Create a function called helpJesse() that accepts an array of Instruction objects as an argument. The function should convert each object into a string of the following format:
"Pour {amount} mL of {solution} into a {instrument}"
It should return an array of each object-to-string conversion, in the same order given.