Simple Fun #320: Scratch lottery I
You got a scratch lottery, you want to know how much money you win.
There are 6 sets of characters on the lottery. Each set of characters represents a chance to win. The text has a coating on it. 
When you buy the lottery ticket and then blow it off, you can see the text information below the coating.
Each set of characters contains three animal names and a number, like this: "tiger tiger tiger 100". 
If the three animal names are the same, Congratulations, you won the prize. You will win the same bonus as the last number.
Given the lottery, returns the total amount of the bonus.